Add page objects and update hooks/config
Introduce BasePage with wait utilities (waitForElement, click, type) using explicit.wait from ConfigReader, and add LoginPage and ProductsPage page objects with common interactions (login fields, add/remove product, cart count). Fix Hooks to parse implicit.wait as Long for Duration.ofSeconds. Update config.properties to include explicit.wait newline and add default username/password.

+++ src/test/java/hooks/Hooks.java
@@ -24,7 +24,7 @@ public class Hooks {
         driver.manage().window().maximize();
         driver.manage().timeouts()
                 .implicitlyWait(Duration.ofSeconds(
-                        Integer.parseInt(ConfigReader.get("implicit.wait"))
+                        Long.parseLong(ConfigReader.get("implicit.wait"))
                 ));
         driver.get(ConfigReader.get("base.url"));
         log.info("Browser launched and navigated to: {}", ConfigReader.get("base.url"));
